refactor(services): extract shared protocol helpers
