+++ src/services/discord/settings.rs
@@ -7,7 +7,7 @@ use sha2::{Digest, Sha256};
 
 use poise::serenity_prelude as serenity;
 
-use crate::services::claude::DEFAULT_ALLOWED_TOOLS;
+use crate::services::agent_protocol::DEFAULT_ALLOWED_TOOLS;
 use crate::services::provider::ProviderKind;
 
 use super::DiscordBotSettings;
